refactor: reorganize quiz template handlers
